# Conflicts:
#	data/src/main/java/com/susu/data/data/repository/EnvelopesRepositoryImpl.kt
#	data/src/main/java/com/susu/data/remote/api/EnvelopesService.kt
#	domain/src/main/java/com/susu/domain/repository/EnvelopesRepository.kt
#	feature/navigator/src/main/java/com/susu/feature/navigator/MainNavigator.kt
#	feature/navigator/src/main/java/com/susu/feature/navigator/MainScreen.kt
#	feature/received/src/main/java/com/susu/feature/received/ledgerdetail/LedgerDetailScreen.kt
#	feature/received/src/main/java/com/susu/feature/received/ledgerdetail/LedgerDetailViewModel.kt
#	feature/received/src/main/java/com/susu/feature/received/navigation/ReceivedNavigation.kt
#	feature/sent/build.gradle.kts
#	feature/sent/src/main/java/com/susu/feature/sent/navigation/SentNavigation.kt
#	feature/sent/src/main/res/values/strings.xml

+++ domain/src/main/java/com/susu/domain/usecase/envelope/GetEnvelopesListUseCase.kt
@@ -21,7 +21,7 @@ class GetEnvelopesListUseCase @Inject constructor(
     }
 
     data class Param(
-        val friendIds: List<Int>? = emptyList(),
+        val friendIds: List<Long>? = emptyList(),
         val fromTotalAmounts: Int? = null,
         val toTotalAmounts: Int? = null,
         val page: Int? = null,
